🔧 refactor(bot, capa, pauta, buscador, wrapper, printlogs): ajustar importações e estrutura de classes
Este commit refatora a estrutura de importações e a lógica de classes em vários arquivos, melhorando a organização do código.

Files Changed:
- *crawjud/abstract/__init__.py*: Adiciona um novo arquivo com uma diretiva de linting.
- *crawjud/bot.py*: Cria o módulo principal de controle dos robôs com a classe `ClassBot`.
- *crawjud/bots/pje/capa.py*: Atualiza a importação da classe `ClassBot`.
- *crawjud/bots/pje/pauta.py*: Atualiza a importação da classe `ClassBot`.
- *crawjud/bots/pje/resources/buscador.py*: Atualiza a importação da classe `ClassBot`.
- *crawjud/wrapper.py*: Atualiza a importação da classe `ClassBot`.
- *utils/printlogs/_master.py*: Atualiza a importação da classe `ClassBot`.
