+++ crawjud/abstract/__init__.py
@@ -1,0 +1,1 @@
+# noqa: D104
