+++ crawjud/bot.py
@@ -17,6 +17,7 @@ from dotenv import dotenv_values
 from pytz import timezone
 from socketio import SimpleClient
 
+from celery_app.tasks.files import SaveSuccessCache
 from utils.models.logs import MessageLogDict
 
 if TYPE_CHECKING:
@@ -150,3 +151,6 @@ class ClassBot(ABC):  # noqa:  D101
                 data["CNPJ_FAVORECIDO"] = "04.812.509/0001-90"
 
         return data
+
+    def save_success_cache(self, data: dict) -> None:  # noqa: D102
+        SaveSuccessCache.apply_async(kwargs=data)
