+++ crawjud/bots/pje/pauta.py
@@ -19,8 +19,8 @@ from selenium.webdriver.support import expected_conditions as ec
 
 from celery_app._wrapper import shared_task
 from celery_app.custom._task import ContextTask
+from crawjud.bot import ClassBot
 from crawjud.bots.pje.resources.varas_dict import varas as varas_pje
-from crawjud.common.bot import ClassBot
 from crawjud.common.exceptions.bot import ExecutionError
